+++ tests/unit/test_chainer.py
@@ -248,6 +248,32 @@ def test_create_model(sagemaker_session, chainer_version):
     assert model.enable_cloudwatch_metrics == enable_cloudwatch_metrics
 
 
+def test_create_model_with_custom_image(sagemaker_session):
+    container_log_level = '"logging.INFO"'
+    source_dir = 's3://mybucket/source'
+    enable_cloudwatch_metrics = 'true'
+    custom_image = 'ubuntu:latest'
+    chainer = Chainer(entry_point=SCRIPT_PATH, role=ROLE, sagemaker_session=sagemaker_session,
+                      train_instance_count=INSTANCE_COUNT, train_instance_type=INSTANCE_TYPE,
+                      image_name=custom_image, container_log_level=container_log_level,
+                      py_version=PYTHON_VERSION, base_job_name='job', source_dir=source_dir,
+                      enable_cloudwatch_metrics=enable_cloudwatch_metrics)
+
+    job_name = 'new_name'
+    chainer.fit(inputs='s3://mybucket/train', job_name='new_name')
+    model = chainer.create_model()
+    chainer.container_log_level
+
+    assert model.sagemaker_session == sagemaker_session
+    assert model.image == custom_image
+    assert model.entry_point == SCRIPT_PATH
+    assert model.role == ROLE
+    assert model.name == job_name
+    assert model.container_log_level == container_log_level
+    assert model.source_dir == source_dir
+    assert model.enable_cloudwatch_metrics == enable_cloudwatch_metrics
+
+
 @patch('time.strftime', return_value=TIMESTAMP)
 def test_chainer(strftime, sagemaker_session, chainer_version):
     chainer = Chainer(entry_point=SCRIPT_PATH, role=ROLE, sagemaker_session=sagemaker_session,
@@ -393,3 +419,46 @@ def test_attach_wrong_framework(sagemaker_session):
     with pytest.raises(ValueError) as error:
         Chainer.attach(training_job_name='neo', sagemaker_session=sagemaker_session)
     assert "didn't use image for requested framework" in str(error)
+
+
+def test_attach_custom_image(sagemaker_session):
+    training_image = '1.dkr.ecr.us-west-2.amazonaws.com/my_custom_chainer_image:latest'
+    returned_job_description = {'AlgorithmSpecification':
+                                {'TrainingInputMode': 'File',
+                                 'TrainingImage': training_image},
+                                'HyperParameters':
+                                    {'sagemaker_submit_directory': '"s3://some/sourcedir.tar.gz"',
+                                     'sagemaker_program': '"iris-dnn-classifier.py"',
+                                     'sagemaker_s3_uri_training': '"sagemaker-3/integ-test-data/tf_iris"',
+                                     'sagemaker_enable_cloudwatch_metrics': 'false',
+                                     'sagemaker_container_log_level': '"logging.INFO"',
+                                     'sagemaker_job_name': '"neo"',
+                                     'training_steps': '100',
+                                     'sagemaker_region': '"us-west-2"'},
+                                'RoleArn': 'arn:aws:iam::366:role/SageMakerRole',
+                                'ResourceConfig':
+                                    {'VolumeSizeInGB': 30,
+                                     'InstanceCount': 1,
+                                     'InstanceType': 'ml.c4.xlarge'},
+                                'StoppingCondition': {'MaxRuntimeInSeconds': 24 * 60 * 60},
+                                'TrainingJobName': 'neo',
+                                'TrainingJobStatus': 'Completed',
+                                'OutputDataConfig': {'KmsKeyId': '',
+                                                     'S3OutputPath': 's3://place/output/neo'},
+                                'TrainingJobOutput': {'S3TrainingJobOutput': 's3://here/output.tar.gz'}}
+    sagemaker_session.sagemaker_client.describe_training_job = Mock(name='describe_training_job',
+                                                                    return_value=returned_job_description)
+
+    estimator = Chainer.attach(training_job_name='neo', sagemaker_session=sagemaker_session)
+    assert estimator.latest_training_job.job_name == 'neo'
+    assert estimator.role == 'arn:aws:iam::366:role/SageMakerRole'
+    assert estimator.train_instance_count == 1
+    assert estimator.train_max_run == 24 * 60 * 60
+    assert estimator.input_mode == 'File'
+    assert estimator.base_job_name == 'neo'
+    assert estimator.output_path == 's3://place/output/neo'
+    assert estimator.output_kms_key == ''
+    assert estimator.hyperparameters()['training_steps'] == '100'
+    assert estimator.source_dir == 's3://some/sourcedir.tar.gz'
+    assert estimator.entry_point == 'iris-dnn-classifier.py'
+    assert estimator.train_image() == training_image